+++ demo/pages/elements/index.ts
@@ -27,3 +27,4 @@ export * from './form/FormDemo';
 export * from './category-dropdown/CategoryDropdownDemo';
 export * from './popover/PopoverDemo';
 export * from './date-picker/DatePickerDemo';
+export * from './field-interactions/FieldInteractionsDemo';