+++ novo-elements.providers.ts
@@ -1,12 +1,15 @@
 // NG2
 import { NgModule } from '@angular/core';
+import { Http } from '@angular/http';
 // APP
 import { NovoDragulaService } from './src/elements/dragula/DragulaService';
 import { NovoModalService } from './src/elements/modal/ModalService';
 import { NovoModalRef } from './src/elements/modal/Modal';
 import { NovoToastService } from './src/elements/toast/ToastService';
 import { ComponentUtils } from './src/utils/component-utils/ComponentUtils';
+import { FormUtils } from './src/utils/form-utils/FormUtils';
 import { DateFormatService } from './src/services/date-format/DateFormat';
+import { FieldInteractionApi } from './src/elements/form/FieldInteractionApi';
 
 const NOVO_ELEMENTS_PROVIDERS = [
     { provide: NovoDragulaService, useClass: NovoDragulaService },
@@ -14,6 +17,7 @@ const NOVO_ELEMENTS_PROVIDERS = [
     { provide: NovoModalService, useClass: NovoModalService },
     { provide: NovoToastService, useClass: NovoToastService },
     { provide: ComponentUtils, useClass: ComponentUtils },
+    FieldInteractionApi,
     DateFormatService
 ];