+++ src/elements/form/NovoFormControl.ts
@@ -15,19 +15,73 @@ export class NovoFormControl extends FormControl {
     tooltip: string;
     tooltipPosition: string;
     initialValue: any;
+    valueHistory: any[] = [];
     validators: any;
+    config: any;
+    sortOrder: number;
+    controlType: string;
+    placeholder: string;
+    multiple: boolean;
+    headerConfig: any;
+    optionsType: string;
+    maxlength: number;
+    minlength: number;
+    options: Array<any>;
+    type: string;
+    subType: string;
+    name: string;
+    closeOnSelect: boolean;
+    interactions: Array<Object>;
+    appendToBody: boolean;
+    parentScrollSelector: string;
+    description?: string;
+    layoutOptions?: { order?: string, download?: boolean, labelStyle?: string, draggable?: boolean, iconStyle?: string };
+    military?: boolean;
+    tipWell?: {
+        tip: string,
+        icon?: string,
+        button?: boolean;
+    };
+
+    private historyTimeout: any;
 
     constructor(value: any, control: NovoControlConfig) {
         super(value, control.validators, control.asyncValidators);
         this.validators = control.validators;
         this.initialValue = value;
+        this.valueHistory.push(value);
         this.label = control.label;
         this.readOnly = control.readOnly;
         this.hidden = control.hidden;
+        this.config = control.config;
+        this.type = control.type;
+        this.subType = control.subType;
         this.required = control.required;
         this.hasRequiredValidator = this.required;
         this.tooltip = control.tooltip;
         this.tooltipPosition = control.tooltipPosition;
+        this.label = control.label;
+        this.name = control.name;
+        this.required = control.required;
+        this.hidden = control.hidden;
+        this.sortOrder = control.sortOrder;
+        this.controlType = control.controlType;
+        this.placeholder = control.placeholder;
+        this.multiple = control.multiple;
+        this.headerConfig = control.headerConfig;
+        this.optionsType = control.optionsType;
+        this.readOnly = control.readOnly;
+        this.layoutOptions = control.layoutOptions;
+        this.military = control.military;
+        this.maxlength = control.maxlength;
+        this.minlength = control.minlength;
+        this.closeOnSelect = control.closeOnSelect;
+        this.interactions = control.interactions;
+        this.appendToBody = control.appendToBody;
+        this.parentScrollSelector = control.parentScrollSelector;
+        this.description = control.description;
+        this.options = control.options;
+        this.tipWell = control.tipWell;
 
         // Reactive Form, need to enable/disable, can't bind to [disabled]
         if (this.readOnly) {
@@ -97,9 +151,14 @@ export class NovoFormControl extends FormControl {
     } = {}) {
         this.markAsDirty();
         this.markAsTouched();
-        // TODO: Should we set defaults on these?
         this.displayValueChanges.emit(value);
         super.setValue(value, { onlySelf, emitEvent, emitModelToViewChange, emitViewToModelChange });
+
+        // History
+        clearTimeout(this.historyTimeout);
+        this.historyTimeout = setTimeout(() => {
+            this.valueHistory.push(value);
+        }, 300);
     }
 
     /**