+++ src/elements/select/Select.ts
@@ -145,18 +145,18 @@ export class NovoSelectElement extends OutsideClick implements OnInit, OnChanges
                     this.saveHeader();
                     return;
                 }
-                this.select(this.options[this.selectedIndex], this.selectedIndex);
+                this.select(this.filteredOptions[this.selectedIndex], this.selectedIndex);
                 this.toggleActive();
                 return;
             }
 
             if (event.keyCode === KeyCodes.UP && this.selectedIndex > 0) {
                 this.selectedIndex--;
-                this.select(this.options[this.selectedIndex], this.selectedIndex);
+                this.select(this.filteredOptions[this.selectedIndex], this.selectedIndex);
                 this.scrollToSelected();
             } else if (event.keyCode === KeyCodes.DOWN && this.selectedIndex < this.options.length - 1) {
                 this.selectedIndex++;
-                this.select(this.options[this.selectedIndex], this.selectedIndex);
+                this.select(this.filteredOptions[this.selectedIndex], this.selectedIndex);
                 this.scrollToSelected();
                 if (this.header.open) {
                     this.toggleHeader(null, false);
@@ -174,9 +174,9 @@ export class NovoSelectElement extends OutsideClick implements OnInit, OnChanges
                 let item = element.querySelector(`[data-automation-value^="${this.filterTerm}" i]`);
                 if (item) {
                     list.scrollTop = item.offsetTop;
-                    let listItems = Array.from(list.querySelectorAll('li')).map((element: any) => element.getAttribute('data-automation-value'));
+                    let listItems = Array.from(list.querySelectorAll('li')).map((e: any) => e.getAttribute('data-automation-value')).filter(value => value);
                     this.selectedIndex = listItems.indexOf(item.getAttribute('data-automation-value'));
-                    this.select(this.options[this.selectedIndex], this.selectedIndex);
+                    this.select(this.filteredOptions[this.selectedIndex], this.selectedIndex);
                 }
             } else if ([KeyCodes.BACKSPACE, KeyCodes.DELETE].includes(event.keyCode)) {
                 clearTimeout(this.filterTermTimeout);